pyproject.toml
[tool.poetry]
name = "stripe-example-api"
version = "0.1.0"
description = ""
authors = ["Maxim Donchuk <m.donchuk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0"
django-rest-framework = "^0.1.0"
stripe = "^7.10.0"
django-environ = "^0.11.2"
ipython = "^8.19.0"
drf-yasg = "^1.21.7"
setuptools = "^69.0.3"
psycopg2-binary = "^2.9.9"
gunicorn = "^21.2.0"
psycopg = "^3.1.16"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[virtualenvs]
create = false
